works

+++ src/raycast.rs
@@ -8,8 +8,14 @@ pub struct RaycastPlugin;
 
 impl Plugin for RaycastPlugin {
     fn build(&self, app: &mut App) {
-        app.add_systems(Update, grapple_hook);
+        app.add_systems(Update, grapple_hook)
+            .init_resource::<GrapplePoint>();
     }
+}
+#[derive(Resource, Default)]
+struct GrapplePoint {
+    point: Vec2,
+    should_draw: bool,
 }
 
 fn grapple_hook(
@@ -27,6 +33,7 @@ fn grapple_hook(
     mut mousebtn_evr: EventReader<MouseButtonInput>,
     mut ext_forces: Query<&mut ExternalForce>,
     transforms: Query<(Entity, &Transform), (Without<PlayerMarker>, Without<MainCamera>)>,
+    mut grapple_point: ResMut<GrapplePoint>,
 ) {
     for mut camera in &mut q_camera {
         camera.translation = ball_query.single().0.translation;
@@ -34,17 +41,26 @@ fn grapple_hook(
 
     //somehow raycast doesn't work perfectly after changing camera position to follow player...
     let ray_dir = mousepos.0;
-    gizmos.line_2d(
-        ball_query.single().0.translation.truncate(),
-        ray_dir,
-        Color::LIME_GREEN,
-    );
-    gizmos.circle_2d(mousepos.0, 0.5, Color::RED);
-    gizmos.line_2d(
-        ball_query.single().0.translation.truncate(),
-        ball_query.single().0.translation.truncate() + ball_query.single().2.linvel,
-        Color::RED,
-    );
+    // gizmos.line_2d(
+    //     ball_query.single().0.translation.truncate(),
+    //     ray_dir,
+    //     Color::LIME_GREEN,
+    // );
+
+    if grapple_point.should_draw {
+        gizmos.line_2d(
+            ball_query.single().0.translation.truncate(),
+            grapple_point.point,
+            Color::BEIGE,
+        );
+    }
+
+    // gizmos.circle_2d(mousepos.0, 0.5, Color::RED);
+    // gizmos.line_2d(
+    //     ball_query.single().0.translation.truncate(),
+    //     ball_query.single().0.translation.truncate() + ball_query.single().2.linvel,
+    //     Color::RED,
+    // );
     for ev in mousebtn_evr.read() {
         match ev.state {
             ButtonState::Pressed => {
@@ -68,7 +84,7 @@ fn grapple_hook(
                             0.5,
                             Vec3::new(mousepos.0.x, mousepos.0.y, 0.0)
                                 .distance(ball_query.single().0.translation)
-                                / 1.25,
+                                / 1.33,
                             // 500.0,
                         ])
                         .local_anchor2(Vec2::ZERO);
@@ -76,13 +92,16 @@ fn grapple_hook(
                     commands
                         .entity(ball_query.single().1)
                         .insert(ImpulseJoint::new(entity, joint));
-
+                    grapple_point.point = mousepos.0;
+                    grapple_point.should_draw = true;
                     //we add a force
-                    let mut velocity_vec = Vec2::ZERO;
+                    let velocity_vec;
 
-                    for (_, _, rb_vels, _) in &mut ball_query {
-                        velocity_vec = rb_vels.linvel;
-                    }
+                    // for (_, _, rb_vels, _) in &mut ball_query {
+                    //     velocity_vec = rb_vels.linvel;
+                    // }
+                    velocity_vec = -ball_query.single().0.translation.truncate() + mousepos.0;
+
                     //GravityScale to 0 for better mechanics
                     *ball_query.single_mut().3 = GravityScale(0.0);
 
@@ -104,8 +123,9 @@ fn grapple_hook(
                     ext_force.force = Vec2::ZERO;
                     ext_force.torque = 0.0;
                 }
+                grapple_point.should_draw = false;
 
-                *ball_query.single_mut().3 = GravityScale(1.0);
+                // *ball_query.single_mut().3 = GravityScale(1.0);
             }
         }
     }
